+++ .github/workflows/build.yml
@@ -104,79 +104,33 @@ jobs:
     needs: decide-refs
     runs-on: ubuntu-latest
     outputs:
-      runner_labels: ${{ steps.resolve-runner.outputs.runner_labels }}
-      runner_group: ${{ steps.resolve-runner.outputs.runner_group }}
+      runner_labels: ${{ steps.initialize-runner.outputs.runner_labels }}
+      runner_group: ${{ steps.initialize-runner.outputs.runner_group }}
     steps:
-      - name: Check out buildroot for runner request script
+      - name: Check out buildroot for runner initialization script
         uses: actions/checkout@v6
         with:
-          sparse-checkout: |
-            request_ephemeral_runner.py
-            resolve_runner_config.py
+          sparse-checkout: initialize_runner.py
           sparse-checkout-cone-mode: false
-      - id: request-unforked-ephemeral-runner
-        name: Request merged-stack ephemeral runner
-        if: ${{ inputs.monorepo-is-fork == false && inputs.force-ephemeral-infra == true }}
-        env:
-          API_URL: ${{ inputs.infra-stage == 'stage-dev' && secrets.EPHEMERAL_RUNNER_UNFORKED_API_URL_DEV || secrets.EPHEMERAL_RUNNER_UNFORKED_API_URL_PROD }}
-          INFRA_STAGE: ${{ inputs.infra-stage }}
-          VARIANT: ${{ needs.decide-refs.outputs.variant }}
-          RUNNER_LABEL: ephemeral-${{ github.run_id }}-${{ github.run_attempt }}
-          MAX_ATTEMPTS: ${{ inputs.runner-request-max-attempts }}
-          RETRY_SECONDS: ${{ inputs.runner-request-retry-seconds }}
-          REQUEST_KIND: unforked
-          MISSING_SECRET_MESSAGE: Missing EPHEMERAL_RUNNER_UNFORKED_API_URL_* secret for infra-stage=${{ inputs.infra-stage }}
-        shell: python
-        run: python3 "${{ github.workspace }}/request_ephemeral_runner.py"
-      - id: request-internal-ephemeral-runner
-        name: Request OT-2 internal channel ephemeral runner
-        if: ${{ inputs.force-ephemeral-infra == true && needs.decide-refs.outputs.variant == 'internal-release' }}
-        env:
-          # ot2-internal-ci → ot2-development.builds.opentrons.com (see .github/workflows/RUNNER_SECRETS.md)
-          API_URL: ${{ inputs.infra-stage == 'stage-dev' && secrets.EPHEMERAL_RUNNER_FORKED_INTERNAL_API_URL_DEV || secrets.EPHEMERAL_RUNNER_FORKED_INTERNAL_API_URL_PROD }}
-          INFRA_STAGE: ${{ inputs.infra-stage }}
-          VARIANT: ${{ needs.decide-refs.outputs.variant }}
-          RUNNER_LABEL: ephemeral-${{ github.run_id }}-${{ github.run_attempt }}
-          MAX_ATTEMPTS: ${{ inputs.runner-request-max-attempts }}
-          RETRY_SECONDS: ${{ inputs.runner-request-retry-seconds }}
-          REQUEST_KIND: ot2-internal
-          MISSING_SECRET_MESSAGE: Missing EPHEMERAL_RUNNER_FORKED_INTERNAL_API_URL_* secret for infra-stage=${{ inputs.infra-stage }}
-        shell: python
-        run: python3 "${{ github.workspace }}/request_ephemeral_runner.py"
-      - id: request-external-ot2-ephemeral-runner
-        name: Request OT-2 external channel ephemeral runner
-        if: ${{ inputs.force-ephemeral-infra == true && inputs.monorepo-is-fork == true && needs.decide-refs.outputs.variant == 'release' }}
-        env:
-          # ot2-ci → ot2.builds.opentrons.com (see .github/workflows/RUNNER_SECRETS.md)
-          API_URL: ${{ inputs.infra-stage == 'stage-dev' && secrets.EPHEMERAL_RUNNER_FORKED_RELEASE_API_URL_DEV || secrets.EPHEMERAL_RUNNER_FORKED_RELEASE_API_URL_PROD }}
-          INFRA_STAGE: ${{ inputs.infra-stage }}
-          VARIANT: ${{ needs.decide-refs.outputs.variant }}
-          RUNNER_LABEL: ephemeral-${{ github.run_id }}-${{ github.run_attempt }}
-          MAX_ATTEMPTS: ${{ inputs.runner-request-max-attempts }}
-          RETRY_SECONDS: ${{ inputs.runner-request-retry-seconds }}
-          REQUEST_KIND: ot2-external
-          MISSING_SECRET_MESSAGE: Missing EPHEMERAL_RUNNER_FORKED_RELEASE_API_URL_* secret for infra-stage=${{ inputs.infra-stage }}
-        shell: python
-        run: python3 "${{ github.workspace }}/request_ephemeral_runner.py"
-      - id: static-runner
-        name: Select merged-stack static runners
-        if: ${{ inputs.monorepo-is-fork == false && inputs.force-ephemeral-infra == false }}
-        run: |
-          echo "runner_labels=[\"self-hosted\", \"${{inputs.infra-stage}}\", \"${{needs.decide-refs.outputs.variant}}\"]" >> "$GITHUB_OUTPUT"
-      - id: resolve-runner
-        name: Resolve runner labels and group
+      - id: initialize-runner
+        name: Initialize runner
         env:
           FORCE_EPHEMERAL: ${{ inputs.force-ephemeral-infra }}
           MONOREPO_IS_FORK: ${{ inputs.monorepo-is-fork }}
           VARIANT: ${{ needs.decide-refs.outputs.variant }}
           INFRA_STAGE: ${{ inputs.infra-stage }}
           MONOREPO_REPO: ${{ needs.decide-refs.outputs.monorepo-repo }}
-          STATIC_LABELS: ${{ steps.static-runner.outputs.runner_labels }}
-          INTERNAL_LABELS: ${{ steps.request-internal-ephemeral-runner.outputs.runner_labels }}
-          EXTERNAL_LABELS: ${{ steps.request-external-ot2-ephemeral-runner.outputs.runner_labels }}
-          UNFORKED_LABELS: ${{ steps.request-unforked-ephemeral-runner.outputs.runner_labels }}
+          RUNNER_LABEL: ephemeral-${{ github.run_id }}-${{ github.run_attempt }}
+          MAX_ATTEMPTS: ${{ inputs.runner-request-max-attempts }}
+          RETRY_SECONDS: ${{ inputs.runner-request-retry-seconds }}
+          EPHEMERAL_RUNNER_FORKED_INTERNAL_API_URL_DEV: ${{ secrets.EPHEMERAL_RUNNER_FORKED_INTERNAL_API_URL_DEV }}
+          EPHEMERAL_RUNNER_FORKED_INTERNAL_API_URL_PROD: ${{ secrets.EPHEMERAL_RUNNER_FORKED_INTERNAL_API_URL_PROD }}
+          EPHEMERAL_RUNNER_FORKED_RELEASE_API_URL_DEV: ${{ secrets.EPHEMERAL_RUNNER_FORKED_RELEASE_API_URL_DEV }}
+          EPHEMERAL_RUNNER_FORKED_RELEASE_API_URL_PROD: ${{ secrets.EPHEMERAL_RUNNER_FORKED_RELEASE_API_URL_PROD }}
+          EPHEMERAL_RUNNER_UNFORKED_API_URL_DEV: ${{ secrets.EPHEMERAL_RUNNER_UNFORKED_API_URL_DEV }}
+          EPHEMERAL_RUNNER_UNFORKED_API_URL_PROD: ${{ secrets.EPHEMERAL_RUNNER_UNFORKED_API_URL_PROD }}
         shell: python
-        run: python3 "${{ github.workspace }}/resolve_runner_config.py"
+        run: python3 "${{ github.workspace }}/initialize_runner.py"
 
   # Build job runs on ephemeral or static runner per initialize-infra
   run-build:
